+++ src/main/resources/application-local.properties
@@ -1,4 +1,4 @@
-spring.application.name=aws-spring-boot-native-dsql
+spring.application.name=cdc
 
 spring.datasource.cluster-endpoint=localhost
 spring.datasource.url=jdbc:postgresql://${spring.datasource.cluster-endpoint}:5432/postgres
