+++ src/main/resources/application.properties
@@ -1,7 +1,5 @@
-spring.application.name=aws-spring-boot-native-dsql
+spring.application.name=cdc
 
-spring.datasource.cluster-endpoint=${CLUSTER_ENDPOINT}
-spring.datasource.url=jdbc:postgresql://${spring.datasource.cluster-endpoint}:5432/postgres
 spring.datasource.username=${DB_USERNAME}
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
