--- src/test/java/com/software/aws_spring_boot_native_dsql/AwsSpringBootNativeDsqlApplicationTests.java
@@ -1,13 +1,0 @@
-package com.software.aws_spring_boot_native_dsql;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AwsSpringBootNativeDsqlApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
